+++ CHANGELOG.md
@@ -8,11 +8,13 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 - Support for Wazuh 4.14.3
 - Added ability to remove agents from Agents management > Summary [#8025](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8025)
+- Added API selector warning in case run_as is false in the wazuh.yml file [#8026](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8026)
 
 ### Changed
 
 - Upgraded `js-yaml` dependency to 4.1.1 [#7958](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7958)
 - Upgraded `swagger-client` dependency to 3.36.0 [#8008](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8008)
+- Set run_as as true by default in the wazuh.yml file [#8026](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8026)
 
 ## Wazuh v4.14.2 - OpenSearch Dashboards 2.19.4 - Revision 02